% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blanc.R
\name{blanc}
\alias{blanc}
\title{Fonction génératrice des cases vides. Elle prend en entrée notre sudoku puis génère
aléatoirement des "trous" à l'aide de la créations de "blocks" temporaires. On fixe une moyenne de cases vide et un nombre maximum de cases vides
par défaut de façon à ce que le sudoku reste bien solvable mais il peut ces données sont modifiables.}
\usage{
blanc(sudo, freq.mean = 1.25, freq.max = 2)
}
\arguments{
\item{sudo}{notre sudoku}

\item{freq.mean}{nombre moyen de blancs}

\item{freq.max}{pour le nombre max de blancs}
}
\value{
Matrice
}
\description{
Fonction génératrice des cases vides. Elle prend en entrée notre sudoku puis génère
aléatoirement des "trous" à l'aide de la créations de "blocks" temporaires. On fixe une moyenne de cases vide et un nombre maximum de cases vides
par défaut de façon à ce que le sudoku reste bien solvable mais il peut ces données sont modifiables.
}
