% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/danmsudo.R
\name{danmsudo}
\alias{danmsudo}
\title{Fonction génératrice de sudoku}
\usage{
danmsudo(seed = NULL, attempt = 1, ultimate = F, retry = T)
}
\arguments{
\item{seed}{aléa création sudoku}

\item{attempt}{compteur}

\item{ultimate}{Pour augmenter le niveau de difficulter , si = F le sudoku sera forcément de niveau facile.}

\item{retry}{On regarde si notre sudoku est solvable}
}
\value{
Matrice
}
\description{
Fonction génératrice de sudoku
}
