% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{solver}
\alias{solver}
\title{Fonction resolution sudoku.}
\usage{
solver(sgrid, mgrid = NULL, blocks = list(), ultimate = F)
}
\arguments{
\item{sgrid}{grille sudoku}

\item{mgrid}{matrice génératrice des blancs}

\item{blocks}{bloques de sudoku sous forme de liste}

\item{ultimate}{Outil pour la fonction solver.sudo}
}
\value{
Matrice
}
\description{
Fonction resolution sudoku.
}
