% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sudo.gaps.R
\name{sudo.gaps}
\alias{sudo.gaps}
\title{Fonction génératrice des cases vides}
\usage{
sudo.gaps(sudo, inference.mean = 1.25, inference.max = 2)
}
\arguments{
\item{sudo}{notre sudoku}

\item{inference.mean}{nombre moyen de blancs}

\item{inference.max}{pour le nombre max de blancs}
}
\value{
Matrice
}
\description{
Fonction génératrice des cases vides
}
