% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/danmsudo.R
\name{danmsudo}
\alias{danmsudo}
\title{Fonction génératrice de sudoku}
\usage{
danmsudo(seed = NULL, attempt = 1, ultimate = F, iterate = T)
}
\arguments{
\item{attempt}{compteur}
}
\value{
Matrice
}
\description{
Fonction génératrice de sudoku
}
