% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/danmsudo.R
\name{danmsudo}
\alias{danmsudo}
\title{Fonction génératrice de sudoku}
\usage{
danmsudo(seed = NULL, attempt = 1, ultimate = F, iterate = T)
}
\arguments{
\item{seed}{aléa création sudoku}

\item{attempt}{compteur}

\item{ultimate}{pour accentuer la teinte foncée de certaines cases}

\item{iterate}{On regarde si notre sudoku est solvable}
}
\value{
Matrice
}
\description{
Fonction génératrice de sudoku
}
