% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{solver}
\alias{solver}
\title{Fonction resolution sudoku}
\usage{
solver(
  sgrid,
  mgrid = NULL,
  noisily = F,
  blocks = list(),
  smatch = 10,
  ultimate = F
)
}
\arguments{
\item{sgrid}{grille sudoku}

\item{mgrid}{.}

\item{noisily}{.}

\item{blocks}{.}

\item{smatch}{.}

\item{ultimate}{.}
}
\value{
Matrice
}
\description{
Fonction resolution sudoku
}
