% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.sudo.R
\name{solver.sudoku}
\alias{solver.sudoku}
\title{Fonction resolution sudoku}
\usage{
solver.sudoku(sudo, ...)
}
\arguments{
\item{sudo}{notre sudoku}
}
\value{
Matrice
}
\description{
Fonction resolution sudoku
}
