% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.sudo.R
\name{solver.sudoku}
\alias{solver.sudoku}
\title{Sous fonction de solver servant à afficher la solution finale}
\usage{
solver.sudoku(sudo, ...)
}
\arguments{
\item{sudo}{notre sudoku}

\item{...}{appel à la fonction solver. voir help(solver)}
}
\value{
Matrice
}
\description{
Sous fonction de solver servant à afficher la solution finale
}
